factorized components
