+++ apps/festival/festival/src/app/marketplace/organization/title/title.component.html
@@ -1,8 +1,8 @@
 <h2>Line-up</h2>
 <section fxLayout="row wrap" fxLayoutGap="24px grid">
   <ng-container *ngFor="let title of titles$ | async; trackBy: trackById">
-    <movie-card [movie]="title" [link]="['../../../title', title.id]" size="poster" [attr.test-id]="'movie-card_' + title.id">
-      <event-request-screening [movieId]="title.id" iconOnly></event-request-screening>
+    <movie-card [movie]="title" [link]="['/c/o/marketplace/title/', title.id]" size="poster" [attr.test-id]="'movie-card_' + title.id" [showWishlistButton]="!isAnonymous">
+      <event-request-screening *ngIf="!isAnonymous" [movieId]="title.id" iconOnly></event-request-screening>
     </movie-card>
   </ng-container>
 </section>
