+++ apps/festival/festival/src/app/marketplace/organization/view/view.component.html
@@ -1,6 +1,6 @@
 <ng-container *ngIf="org$ | async as org">
   <org-view [org]="org" [navLinks]="navLinks">
-    <ng-container *ngIf="org.email">
+    <ng-container *ngIf="org.email && !isAnonymous">
       <a mat-flat-button color="accent"
         [href]="'mailto:' + org.email + '?subject=Meeting%20request%20via%20Archipel%20Market'">
         <span>Ask for a meeting</span>
