--- apps/festival/festival/src/app/organization/view/view.component.html
@@ -1,10 +1,0 @@
-<ng-container *ngIf="org$ | async as org">
-  <org-view [org]="org" [navLinks]="navLinks">
-    <ng-container *ngIf="org.email && !isAnonymous">
-      <a mat-flat-button color="accent"
-        [href]="'mailto:' + org.email + '?subject=Meeting%20request%20via%20Archipel%20Market'">
-        <span>Ask for a meeting</span>
-      </a>
-    </ng-container>
-  </org-view>
-</ng-container>
