--- apps/festival/festival/src/app/organization/view/view.component.scss
@@ -1,3 +1,0 @@
-:host {
-  display: block;
-}
